requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
pyyaml>=6.0
numpy>=1.24.0
matplotlib>=3.8.0
pandas>=2.0.0
Pillow>=10.0.0
pycocotools>=2.0.7
tqdm>=4.66.0
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
